fix script test/pdv/pdvPerformance.js
