+++ test/common/loadEnv.js
@@ -1,17 +1,33 @@
-import { toTrimmedString } from './utils.js'
+import { toTrimmedString } from './basicUtils.js'
 
 export function loadEnvConfig(env) {
     const normalizedEnv = toTrimmedString(env, 'dev').toLowerCase()
-    const path = `../../config/${normalizedEnv}.json`
+    const resolvedPath = import.meta.resolve(`../../config/${normalizedEnv}.json`)
+    const filePath = resolvedPath.startsWith('file://')
+        ? normalizeFileUrl(resolvedPath)
+        : resolvedPath
 
     try {
-        const raw = open(path)
+        const raw = open(filePath)
         const parsed = JSON.parse(raw)
         if (!parsed || typeof parsed.pdvUrl !== 'string') {
-            throw new Error(`config file ${path} missing pdvUrl property`)
+            throw new Error(
+                `config file ${filePath} missing pdvUrl property`
+            )
         }
         return parsed
     } catch (err) {
-        throw new Error(`Unable to load config for TARGET_ENV=${normalizedEnv}: ${err.message}`)
+        throw new Error(
+            `Unable to load config for TARGET_ENV=${normalizedEnv}: ${err.message}`
+        )
     }
+}
+
+function normalizeFileUrl(fileUrl) {
+    let path = fileUrl.slice('file://'.length)
+    // k6 returns paths as file:///C:/... on Windows; drop the extra slash
+    if (path.startsWith('/') && /^[A-Za-z]:/.test(path.slice(1))) {
+        return path.slice(1)
+    }
+    return path
 }
